+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+## [0.8.0] - 2017-03-11
+### Fixed
+- Move pry-byebug require to `spec_helper`
+
 ## [0.7.0] - 2017-02-01
 ### Added
 - Set secret from file